Fixed `end` event incompatibility with any-db spec
Now `end` is not emitted if there was no callback or `data` event
handler set on query. This makes adapter pass 56/58 any-db adapter spec
tests. The last two are about missing `pause()` and `resume()` methods,
which should come from `stream.Readable`, but Tedious Request is not a
stream, so there's no clean way to implement that (except for
pseudo-stream emulation, which would be kinda ugly and could confuse
developers into thinking that query can be paused, without taking up
memory for all rows).